1.8_AI DRIVEN/1.8-problmstmnt-main (2) 2109224/1.8-problmstmnt-main/requirements.txt
Flask: For the web framework.
Werkzeug: For handling file uploads securely.
pytesseract: For OCR text extraction from images.
Pillow: For opening and handling image files.
pdfplumber: For extracting text from PDF files.
Whoosh: For indexing and searching the text.
re (Regular Expressions): This is part of Python’s standard library, so it does not need to be included.


Flask==2.3.2
Werkzeug==2.3.2
pytesseract==0.3.10
Pillow==9.2.0
pdfplumber==0.7.4
Whoosh==2.7.4
